+++ spring-batch-core/src/main/java/org/springframework/batch/core/job/flow/JobFlowExecutor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2018 the original author or authors.
+ * Copyright 2006-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -84,7 +84,7 @@ public class JobFlowExecutor implements FlowExecutor {
 	}
 
 	private boolean isStepRestart(Step step) {
-		int count = jobRepository.getStepExecutionCount(execution.getJobInstance(), step.getName());
+		long count = jobRepository.getStepExecutionCount(execution.getJobInstance(), step.getName());
 
 		return count > 0;
 	}
