+++ spring-batch-core/src/test/java/org/springframework/batch/core/configuration/xml/DummyJobRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2021 the original author or authors.
+ * Copyright 2006-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -70,7 +70,7 @@ public class DummyJobRepository implements JobRepository, BeanNameAware {
 	}
 
 	@Override
-	public int getStepExecutionCount(JobInstance jobInstance, String stepName) {
+	public long getStepExecutionCount(JobInstance jobInstance, String stepName) {
 		return 0;
 	}
