+++ spring-batch-core/src/test/java/org/springframework/batch/core/step/item/TaskletStepExceptionTests.java
@@ -527,7 +527,7 @@ class TaskletStepExceptionTests {
 		}
 
 		@Override
-		public int getStepExecutionCount(JobInstance jobInstance, String stepName) {
+		public long getStepExecutionCount(JobInstance jobInstance, String stepName) {
 			return 0;
 		}
